--- .jshintrc
@@ -1,104 +1,0 @@
-{
-    "predef": [
-        "confirm",
-        "console",
-        "indexedDB",
-        "Promise",
-        "Map",
-        "Set",
-        "IDBKeyRange",
-        "_gaq",
-        "ZeroClipboard",
-        "dragula",
-        "Bloodhound",
-        "d3",
-        "html2canvas",
-        "Newick",
-        "requestAnimationFrame",
-        "self",
-        "get",
-        "getJSON",
-        "init_phylogram",
-        "constructSearchtree",
-        "constructGenomeTable",
-        "constructGenomeSelector",
-        "constructPancoreGraph",
-        "constructSimMatrix",
-        "constructSunburst",
-        "constructTreemap",
-        "constructMyGenomes",
-        "constructTreeview",
-        "downloadDataByForm",
-        "showNotification",
-        "delay",
-        "stringHash",
-        "logToGoogle",
-        "info",
-        "highlight",
-        "error",
-        "addCopy",
-        "iteratorToArray",
-        "fullScreenApi",
-        "triggerDownloadModal",
-        "getReadableColorFor"
-    ],
-
-    /*** Enforcing options ***/
-    // Set these to `true` to enforce, or `false` to relax.
-    "bitwise": false,
-    "curly": true,
-    "eqeqeq": true,
-    "forin": true,
-    "immed": true,
-    "latedef": false,
-    "newcap": true,
-    "noarg": true,
-    "noempty": true,
-    "nonew": true,
-    "plusplus": false,
-    "regexp": false,
-    "undef": true,
-    "strict": false,
-    "trailing": true,
-
-    /*** Relaxing options ***/
-    // Set these to `true` to relax, or `false` to enforce.
-    "asi": false,
-    "boss": false,
-    "debug": false,
-    "eqnull": true,
-    "es5": false,
-    "esnext": false,
-    "evil": false,
-    "expr": false,
-    "funcscope": false,
-    "globalstrict": false,
-    "iterator": false,
-    "lastsemic": false,
-    "laxbreak": false,
-    "laxcomma": false,
-    "loopfunc": true,
-    "multistr": false,
-    "onecase": false,
-    "proto": false,
-    "regexdash": false,
-    "scripturl": false,
-    "shadow": false,
-    "smarttabs": false,
-    "sub": false,
-    "supernew": false,
-    "validthis": false,
-
-
-    /*** Legacy from JSLint ***/
-    // Set these to `true` to enforce, or `false` to relax.
-    "nomen": false,
-    "onevar": false,
-    "passfail": false,
-    "white": false,
-
-    "browser": true,
-    "jquery": true,
-
-    "esversion": 6
-}
