+++ app/assets/javascripts/polyfills.js
@@ -5,31 +5,33 @@
  * heavily adapted to support IE11 and the new specs
  */
 (function () {
-    var fullScreenApi = {
+    let fullScreenApi = {
             supportsFullScreen: false,
-            isFullScreen: function () { return false; },
+            isFullScreen: function () {
+                return false;
+            },
             requestFullScreen: function () {},
             cancelFullScreen: function () {},
-            fullScreenEventName: '',
-            prefix: ''
+            fullScreenEventName: "",
+            prefix: "",
         },
-        browserPrefixes = 'webkit moz o ms khtml'.split(' '),
+        browserPrefixes = "webkit moz o ms khtml".split(" "),
         i,
         il;
 
     // check for native support
-    if (typeof document.exitFullScreen !== 'undefined') {
+    if (typeof document.exitFullScreen !== "undefined") {
         fullScreenApi.supportsFullScreen = true;
     } else {
         // check for fullscreen support by vendor prefix
         for (i = 0, il = browserPrefixes.length; i < il; i++) {
             fullScreenApi.prefix = browserPrefixes[i];
 
-            if (typeof document[fullScreenApi.prefix + 'CancelFullScreen'] !== 'undefined') {
+            if (typeof document[fullScreenApi.prefix + "CancelFullScreen"] !== "undefined") {
                 fullScreenApi.supportsFullScreen = true;
                 break;
             }
-            if (typeof document[fullScreenApi.prefix + 'ExitFullscreen'] !== 'undefined') {
+            if (typeof document[fullScreenApi.prefix + "ExitFullscreen"] !== "undefined") {
                 fullScreenApi.supportsFullScreen = true;
                 break;
             }
@@ -38,54 +40,54 @@
 
     // update methods to do something useful
     if (fullScreenApi.supportsFullScreen) {
-        fullScreenApi.fullScreenEventName = fullScreenApi.prefix + 'fullscreenchange';
+        fullScreenApi.fullScreenEventName = fullScreenApi.prefix + "fullscreenchange";
         if (fullScreenApi.prefix === "ms") {
             fullScreenApi.fullScreenEventName = "MSFullscreenChange";
         }
 
         fullScreenApi.isFullScreen = function () {
             switch (this.prefix) {
-            case '':
+            case "":
                 return document.fullscreenElement !== null;
-            case 'moz':
+            case "moz":
                 return document.mozFullScreenElement !== null;
             default:
-                return document[this.prefix + 'FullscreenElement'] !== null;
+                return document[this.prefix + "FullscreenElement"] !== null;
             }
         };
         fullScreenApi.requestFullScreen = function (el) {
             switch (this.prefix) {
-            case '':
+            case "":
                 return el.requestFullscreen();
-            case 'webkit':
+            case "webkit":
                 return el.webkitRequestFullscreen();
-            case 'ms':
+            case "ms":
                 return el.msRequestFullscreen();
-            case 'moz':
+            case "moz":
                 return el.mozRequestFullScreen();
-            case 'default':
+            case "default":
                 return el[this.prefix + "RequestFullscreen"]();
             }
         };
         fullScreenApi.cancelFullScreen = function (el) {
             switch (this.prefix) {
-            case '':
+            case "":
                 return document.exitFullscreen();
-            case 'webkit':
+            case "webkit":
                 return document.webkitExitFullscreen();
-            case 'ms':
+            case "ms":
                 return document.msExitFullscreen();
-            case 'moz':
+            case "moz":
                 return document.mozCancelFullScreen();
-            case 'default':
+            case "default":
                 return document[this.prefix + "ExitFullscreen"]();
             }
         };
     }
 
     // jQuery plugin
-    if (typeof jQuery !== 'undefined') {
-        jQuery.fn.requestFullScreen = function () {
+    if (typeof $ !== "undefined") {
+        $.fn.requestFullScreen = function () {
             return this.each(function () {
                 if (fullScreenApi.supportsFullScreen) {
                     fullScreenApi.requestFullScreen(this);
